chore: add client cache
